save changes

+++ deploy.sh
@@ -2,16 +2,17 @@
 set -e
 
 # Check if required arguments are provided
-if [ "$#" -ne 2 ]; then
-    echo "Usage: $0 <server_user> <server_ip>"
+if [ "$#" -ne 1 ]; then
+    echo "Usage: $0 <user@ip>"
     exit 1
 fi
 
 # Parameters
-SERVER_USER="$1"  # SSH user for the server (passed as the first argument)
-SERVER_IP="$2"    # Server IP or hostname (passed as the second argument)
+SERVER="$1"  # SSH user for the server (passed as the first argument)
 SERVER_DIR="~/"  # Target directory on the server
-TAR_FILE="deployment_bundle.tar"  # Tar file name
+
+PROJECT_NAME="word-games"
+TAR_FILE="$PROJECT_NAME-bundle.tar"  # Tar file name
 
 # Paths to files
 DOCKER_COMPOSE_FILE="docker-compose.yml"  # Path to docker-compose.yml
@@ -32,41 +33,46 @@ for IMAGE in "${IMAGES[@]}"; do
     rm "$(basename "$IMAGE").tar"
 done
 
+INSTALL_DIR=~/deployments/$PROJECT_NAME
+# ??
+# ssh "$SERVER_USER@$SERVER_IP" "mkdir -p $INSTALL_DIR"
+
 # Step 3: Transfer tarball to the server
 echo "Transferring tarball to the server..."
-scp "$TAR_FILE" "$SERVER_USER@$SERVER_IP:$SERVER_DIR"
+scp "$TAR_FILE" "$SERVER:$INSTALL_DIR"
 
-
+rm "$TAR_FILE"
 
 # Step 4: SSH into the server and handle deployment
 echo "Deploying on the server..."
 ssh "$SERVER_USER@$SERVER_IP" << EOF
     # Go to the deployment directory
-    cd $SERVER_DIR
+    cd $INSTALL_DIR
+
+    if [ -f "docker-compose.yml" ] || [ -f "compose.yaml" ]; then
+        echo "Stopping containers and removing previous images"
+        docker compose down --rmi all
+    else
+        echo "No Compose file found. Skipping 'docker compose down'."
+    fi
 
     # Untar the new deployment bundle
     echo "Extracting deployment bundle..."
     tar -xf "$TAR_FILE"
 
-    # # Load Docker images
+    # Load Docker images
     echo "Loading Docker images..."
     for IMAGE_TAR in *.tar; do
+        echo "Loading $IMAGE_TAR"
         docker load -i "\$IMAGE_TAR"
     done
 
-    # # # Start the application with Docker Compose
-    # echo "Starting the application..."
-    docker compose down --rmi all
-
+    echo "Starting the application..."
     docker compose up -d
 
-    # # Clean up tar files
+    # Clean up tar files
     echo "Cleaning up tar files..."
     rm -f *.tar
-
-    # Clean up previous deployment
-    # echo "Cleaning up old deployment..."
-    # rm -rf docker-compose.yml *.tar
 EOF
 
 # Step 5: Clean up local tar files
